fixed gobs of warnings
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@44198 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: ccf1ee462817e92f41310fe7ff8b34eab78e5d3c

+++ rosetta_source/src/apps/pilot/will/genI213_nocontacts.cc
@@ -372,8 +372,8 @@ sicfast(
   ymx = min(ymx,ymx1); ymn = max(ymn,ymn1);
 
 
-  int xlb = floor(xmn/BIN)-2; int xub = ceil(xmx/BIN)+2; // one extra on each side for correctness,
-  int ylb = floor(ymn/BIN)-2; int yub = ceil(ymx/BIN)+2; // and one extra for outside atoms
+  int xlb = (int)floor(xmn/BIN)-2; int xub = (int)ceil(xmx/BIN)+2; // one extra on each side for correctness,
+  int ylb = (int)floor(ymn/BIN)-2; int yub = (int)ceil(ymx/BIN)+2; // and one extra for outside atoms
 
   // TR << "BOUNDS " << xmn << " " << xmx << " " << ymn << " " << ymx << std::endl;
   // TR << "BOUNDS " << xlb << " " << xub << " " << ylb << " " << yub << std::endl;
